Move some validators to separate classes
