+++ app/Model/Action.php
@@ -2,9 +2,6 @@
 
 namespace Kanboard\Model;
 
-use SimpleValidator\Validator;
-use SimpleValidator\Validators;
-
 /**
  * Action Model
  *
@@ -187,28 +184,5 @@ class Action extends Base
         }
 
         return true;
-    }
-
-    /**
-     * Validate action creation
-     *
-     * @access public
-     * @param  array   $values           Required parameters to save an action
-     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
-     */
-    public function validateCreation(array $values)
-    {
-        $v = new Validator($values, array(
-            new Validators\Required('project_id', t('The project id is required')),
-            new Validators\Integer('project_id', t('This value must be an integer')),
-            new Validators\Required('event_name', t('This value is required')),
-            new Validators\Required('action_name', t('This value is required')),
-            new Validators\Required('params', t('This value is required')),
-        ));
-
-        return array(
-            $v->execute(),
-            $v->getErrors()
-        );
     }
 }
