+++ app/Validator/ActionValidator.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace Kanboard\Validator;
+
+use SimpleValidator\Validator;
+use SimpleValidator\Validators;
+
+/**
+ * Action Validator
+ *
+ * @package  validator
+ * @author   Frederic Guillot
+ */
+class ActionValidator extends Base
+{
+    /**
+     * Validate action creation
+     *
+     * @access public
+     * @param  array   $values           Required parameters to save an action
+     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
+     */
+    public function validateCreation(array $values)
+    {
+        $v = new Validator($values, array(
+            new Validators\Required('project_id', t('The project id is required')),
+            new Validators\Integer('project_id', t('This value must be an integer')),
+            new Validators\Required('event_name', t('This value is required')),
+            new Validators\Required('action_name', t('This value is required')),
+            new Validators\Required('params', t('This value is required')),
+        ));
+
+        return array(
+            $v->execute(),
+            $v->getErrors()
+        );
+    }
+}
