+++ app/Validator/ColumnValidator.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace Kanboard\Validator;
+
+use SimpleValidator\Validator;
+use SimpleValidator\Validators;
+
+/**
+ * Column Validator
+ *
+ * @package  validator
+ * @author   Frederic Guillot
+ */
+class ColumnValidator extends Base
+{
+    /**
+     * Validate column modification
+     *
+     * @access public
+     * @param  array   $values           Required parameters to update a column
+     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
+     */
+    public function validateModification(array $values)
+    {
+        $v = new Validator($values, array(
+            new Validators\Integer('task_limit', t('This value must be an integer')),
+            new Validators\Required('title', t('The title is required')),
+            new Validators\MaxLength('title', t('The maximum length is %d characters', 50), 50),
+        ));
+
+        return array(
+            $v->execute(),
+            $v->getErrors()
+        );
+    }
+
+    /**
+     * Validate column creation
+     *
+     * @access public
+     * @param  array   $values           Required parameters to save an action
+     * @return array   $valid, $errors   [0] = Success or not, [1] = List of errors
+     */
+    public function validateCreation(array $values)
+    {
+        $v = new Validator($values, array(
+            new Validators\Required('project_id', t('The project id is required')),
+            new Validators\Integer('project_id', t('This value must be an integer')),
+            new Validators\Required('title', t('The title is required')),
+            new Validators\MaxLength('title', t('The maximum length is %d characters', 50), 50),
+        ));
+
+        return array(
+            $v->execute(),
+            $v->getErrors()
+        );
+    }
+}
